Create prediction specific routes for prediction endpoints
Also updated the tests
