+++ src/controllers/measure.controller.js
@@ -36,37 +36,6 @@ const getTrends = async (req, res, next) => {
   }
 };
 
-// Access predictions made by time series
-const getPredictions = async (req, res, next) => {
-  try {
-    const predictions = await dao.findPredictions();
-    return res.send(predictions);
-  } catch (e) {
-    return next(e);
-  }
-};
-
-// Data time series to make predictions with
-const getPredictionData = async (req, res, next) => {
-  try {
-    const search = await dao.findMeasureResults(req.params);
-    const predictionData = search.sort((a, b) => a.date - b.date);
-    const compiledData = {
-      _id: req.params.measure,
-      DATE: {},
-      HEDIS0: {},
-    };
-    for (let i = 0; i < predictionData.length; i += 1) {
-      const result = predictionData[i];
-      compiledData.DATE[i.toString()] = new Date(result.date).getTime();
-      compiledData.HEDIS0[i.toString()] = result.value;
-    }
-    return res.send([compiledData]);
-  } catch (e) {
-    return next(e);
-  }
-};
-
 // Compiles individual info records into one JSON object
 const getInfo = async (req, res, next) => {
   try {
@@ -111,15 +80,6 @@ const postMeasureResults = async (req, res, next) => {
   }
 };
 
-const postPredictions = async (req, res, next) => {
-  try {
-    const predictions = await dao.insertPredictions(req.body);
-    return res.send(predictions);
-  } catch (e) {
-    return next(e);
-  }
-};
-
 const postInfo = async (req, res, next) => {
   try {
     const info = await dao.insertInfo(req.body);
@@ -133,12 +93,9 @@ module.exports = {
   getMeasures,
   getMeasureResults,
   getTrends,
-  getPredictions,
-  getPredictionData,
   getInfo,
   postBulkMeasures,
   postMeasure,
   postMeasureResults,
-  postPredictions,
   postInfo,
 };
