+++ test/controllers/fail.measure.controller.test.js
@@ -7,12 +7,9 @@ const {
   getMeasures,
   getMeasureResults,
   getTrends,
-  getPredictions,
-  getPredictionData,
   postBulkMeasures,
   postMeasure,
   postMeasureResults,
-  postPredictions,
 } = require('../../src/controllers/measure.controller');
 
 const data = JSON.parse(fs.readFileSync(`${path.resolve()}/test/resources/bulk-data.json`));
@@ -73,22 +70,6 @@ describe('## measure.controller.js exceptions', () => {
     });
   });
 
-  describe('Test getPredictions', () => {
-    it('Should catch error and call next', async () => {
-      const next = jest.fn();
-      await getPredictions({ }, jest.fn(), next);
-      expect(next).toHaveBeenCalled();
-    });
-  });
-
-  describe('Test getPredictionData', () => {
-    it('Should catch error and call next', async () => {
-      const next = jest.fn();
-      await getPredictionData({ body: data }, jest.fn(), next);
-      expect(next).toHaveBeenCalled();
-    });
-  });
-
   describe('Test postBulkMeasures upload', () => {
     it('Should catch error and call next', async () => {
       const next = jest.fn();
@@ -109,14 +90,6 @@ describe('## measure.controller.js exceptions', () => {
     it('Should catch error and call next', async () => {
       const next = jest.fn();
       await postMeasureResults({ body: data }, jest.fn(), next);
-      expect(next).toHaveBeenCalled();
-    });
-  });
-
-  describe('Test postPredictions', () => {
-    it('Should catch error and call next', async () => {
-      const next = jest.fn();
-      await postPredictions({ body: data }, jest.fn(), next);
       expect(next).toHaveBeenCalled();
     });
   });
